pyproject.toml
[tool.poetry]
name = "pypicamp"
version = "0.1.0"
description = "PyPI Cache helper for Python Argentina's PyCamp"
authors = ["lecovi <lecovi@ac.python.org.ar>"]
license = "GPL 3"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^12.0.0"
netifaces = "^0.11.0"
requests = "^2.27.1"
speedtest-cli = "^2.1.3"
sendgrid = "^6.9.7"
python-dotenv = "^0.19.2"
typer = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
